Prevent connection failures from stopping async polling
This patch fixes an issue where an async action might make the system
unreachable for a period of time (ie. during a reboot or network restart).

In addition to the above, this patch uses the retry message callback to
provide status updates of the async polling, though it may require some
additional tweaks to the default callback to prevent the messages from
appearing as failures rather than "job not finished" messages.